pyproject.toml
[tool.poetry]
name = "viamagus-chat-app"
version = "0.1.0"
description = ""
authors = ["Arunprakaash <62831206+Arunprakaash@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pinecone-client = "^4.1.1"
openai = "^1.34.0"
python-dotenv = "^1.0.1"
chainlit = "^1.1.302"
sentence-transformers = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
